requirements.txt
streamlit
python-dotenv
groq
chromadb
sentence-transformers
langchain
langchain-community
langchain-text-splitters
numpy
pandas
scikit-learn
